Initial Commit
Contains a bunch of helper scripts and an example attempting to showcase this package.

The repository is at a more advanced state than it usually is for an initial commit.
This is because this code has been transferred from a different project of mine, that was never mean tto get this far.